refactor: move pa cutting and nesting functionlities into different clients with their own interfaces
